Clean up KunstenaarsbrievenBlock components
