+++ src/projects/kunstenaarsbrieven/annotation/block/findBlockSegments.tsx
@@ -1,0 +1,7 @@
+import { Element, Segment } from "../../../../components/Text/Annotated/core";
+
+export function findBlockSegments(elements: Element[]): Segment[] {
+  return elements.flatMap((e) =>
+    e.isBlock ? findBlockSegments(e.children) : e.segments,
+  );
+}
