+++ src/projects/kunstenaarsbrieven/config/index.tsx
@@ -33,7 +33,7 @@ import { TocPanel } from "../TocPanel.tsx";
 import { getUrl, isLink } from "../annotation/LinkUtils.ts";
 import { filterPanels } from "../filterPanels.ts";
 
-import { KunstenaarsbrievenBlock } from "../annotation/KunstenaarsbrievenBlock.tsx";
+import { KunstenaarsbrievenBlock } from "../annotation/block/KunstenaarsbrievenBlock.tsx";
 import { KunstenaarsbrievenHighlight } from "../annotation/KunstenaarsbrievenHighlight.tsx";
 
 export const kunstenaarsbrievenConfig: DefaultProjectConfig = merge(
